refact: move dependencies into separate module; move function that should not be directly used into a utils package
