+++ taf/api/roles.py
@@ -8,7 +8,7 @@ import json
 from pathlib import Path
 from logdecorator import log_on_end, log_on_error, log_on_start
 from tuf.repository_tool import Repository as TUFRepository, Targets
-from taf.api.utils import check_if_clean, commit_and_push
+from taf.api.utils.git_utils import check_if_clean, commit_and_push
 from taf.exceptions import TAFError
 from taf.models.converter import from_dict
 from taf.models.types import RolesIterator
@@ -24,7 +24,10 @@ from taf.keys import (
     load_signing_keys,
     load_sorted_keys_of_new_roles,
 )
-from taf.api.metadata import update_snapshot_and_timestamp, update_target_metadata
+from taf.api.utils.metadata_utils import (
+    update_snapshot_and_timestamp,
+    update_target_metadata,
+)
 from taf.auth_repo import AuthenticationRepository
 from taf.constants import (
     DEFAULT_ROLE_SETUP_PARAMS,
