+++ taf/tests/test_api/test_metadata.py
@@ -8,11 +8,7 @@ from pytest import fixture
 from taf.api.repository import create_repository
 from taf.tests.conftest import CLIENT_DIR_PATH
 from taf.utils import on_rm_error
-from taf.api.metadata import (
-    check_expiration_dates,
-    update_metadata_expiration_date,
-    update_snapshot_and_timestamp,
-)
+from taf.api.metadata import check_expiration_dates, update_metadata_expiration_date
 
 
 AUTH_REPO_NAME = "auth"
@@ -143,20 +139,6 @@ def test_check_expiration_date_when_no_expired(auth_repo_path):
     )
     assert not len(expired)
     assert not len(will_expire)
-
-
-@freeze_time("2023-01-01")
-def test_update_snapshot_and_timestamp(auth_repo_path, api_keystore):
-    auth_repo = AuthenticationRepository(path=auth_repo_path)
-    # signs snapshot and timestamp, uses default expiration intervals
-    update_snapshot_and_timestamp(
-        auth_repo,
-        keystore=api_keystore,
-    )
-    now = datetime.datetime.now()
-    for role, interval in [("timestamp", 1), ("snapshot", 7)]:
-        actual_expiration = auth_repo.get_expiration_date(role)
-        assert now + datetime.timedelta(interval) == actual_expiration
 
 
 def _check_expired_role(role_name, start_time, interval, expired_dict):
